Time: 0 ms (100%), Space: 27.7 MB (41%) - LeetHub

+++ 3637-trionic-array-i/3637-trionic-array-i.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isTrionic(vector<int>& nums) {
         if(nums.size()==3) return false;
-        int state = 0;
+        short state = 0;
         for(int i = 1; i<nums.size(); ++i) {
             if (nums[i] == nums[i-1]) return false;
             if(state==0) //init
